requirements.txt
exa_py==1.0.9
#langsmith==0.1.59
duckduckgo-search
youtube-transcript-api
youtube-dl
python-dateutil
colorama
numexpr


#[tool.poetry.dependencies]
#python">=3.10,<=3.13
langchain==0.1.10
openai==1.13.3
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
regex==2023.12.25
click==8.1.7
python-dotenv==1.0.0
embedchain==0.1.98
appdirs==1.4.4

#[tool.poetry.group.dev.dependencies]
isort==5.13.2
mypy==1.10.0
autoflake==2.2.1
pre-commit==3.6.0
mkdocs==1.4.3
mkdocstrings==0.25.0
mkdocstrings-python==1.1.2
mkdocs-material<9.4
mkdocs-material[imaging]<9.4
mkdocs-material-extensions==1.3.1
pillow==10.2.0
cairosvg==2.7.1
crewai-tools==0.2.6

#[tool.poetry.group.test.dependencies]
pytest==8.0.0
pytest-vcr==1.0.2
python-dotenv==1.0.0


pydantic<3.0.0
instructor==0.5.2
crewai-tools==0.2.6
langchain_openai<0.0.6,>=0.0.5
langchain-community<0.2.0
crewai==0.30.11
poetry==1.8.3%   


google-search-results==2.4.2

